Fix clang warning.

+++ CesiumGltf/include/CesiumGltf/Model.h
@@ -50,7 +50,7 @@ namespace CesiumGltf {
             if (index < 0 || static_cast<size_t>(index) >= pItems->size()) {
                 return nullptr;
             } else {
-                return &(*pItems)[index];
+                return &(*pItems)[size_t(index)];
             }
         }
